requirements.txt
# Core dependencies
fastapi
pillow
numpy
torch
torchvision

# Optional: Upscaling backends
py-real-esrgan
super-image
basicsr>=1.4.2

# Optional: Metrics tools (install for quality metrics)
# If installation fails, metrics will be disabled gracefully
pyiqa>=0.1.0
scikit-image>=0.19.0

